removed league stats from game_manager and implemented methods in game_team_manager
